requirements.txt
certifi==2018.1.18
chardet==3.0.4
idna==2.6
Khayyam==3.0.17
numpy==1.14.2
packaging==17.1
pandas==0.22.0
pyparsing==2.2.0
python-dateutil==2.7.2
pytz==2018.4
requests==2.18.4
six==1.11.0
urllib3==1.22